TypeScript Interfaces to Classes
I didn't realize that interfaces do not automatically declare variables
thus have converted the IUser, and IUserFlags into classes and made them
more useful long term.

* Added User Class
* Changed Flags into a number so that it can be used for session based
flags.
* Added Permissions class to replace the old flags, with two properties
and a method. global is for global permissions flags, channel for
local/channel level ones, and check() determines if a person has
sufficient permissions. Plan to add a checkMany() or change check() to
deal with an array of numbers.